Refactor: Move LinkUI components to dedicated /link-ui directory
- Move link-ui.js to link-ui/index.js
- Move block-inserter.js to link-ui/
- Move page-creator.js to link-ui/
- Move dialog-wrapper.js to link-ui/
- Create shared/index.js to export LinkUI
- Update all import statements to use new paths
- Preserve file history using git mv

This refactoring improves code organization by grouping all LinkUI-related
components in a dedicated directory while maintaining a clean shared API.
