+++ packages/block-library/src/navigation/edit/menu-inspector-controls.js
@@ -23,8 +23,7 @@ import { unlock } from '../../lock-unlock';
 import DeletedNavigationWarning from './deleted-navigation-warning';
 import useNavigationMenu from '../use-navigation-menu';
 import LeafMoreMenu from './leaf-more-menu';
-import { updateAttributes } from '../../navigation-link/shared';
-import { LinkUI } from '../../navigation-link/link-ui';
+import { LinkUI, updateAttributes } from '../../navigation-link/shared';
 
 const actionLabel =
 	/* translators: %s: The name of a menu. */ __( "Switch to '%s'" );
